Criado diretorio Layout, fragmento assim não sobrecarregar as  view
