+++ front-end-senior/desafio-ident/src/layouts/index.js
@@ -1,2 +1,3 @@
 export { default as CarouselLancamentos } from './CarouselLancamentos'
 export { default as Implantodontia } from './Implantodontia'
+export { default as Periodontia } from './Periodontia'
