+++ front-end-senior/desafio-ident/src/views/Home/Home.jsx
@@ -1,8 +1,5 @@
 import React, { useState } from "react";
-import { Panel, CarouselCard, ContentLoading } from "components";
-import { CarouselLancamentos, Implantodontia } from "layouts";
-import imageFile from "../../assets/img1.png";
-import Grid from "@material-ui/core/Grid";
+import { CarouselLancamentos, Implantodontia, Periodontia } from "layouts";
 import { withStyles } from "@material-ui/core/styles";
 
 const styles = theme => ({
@@ -17,70 +14,11 @@ const styles = theme => ({
   }
 });
 
-const Loading = () => {
-  return Array(4).fill(4).map((e,i) => (
-    <Grid key={i} item xs={12} md={6} lg={3}>
-      <ContentLoading />
-    </Grid>
-  ));
-};
-
-/*const Implantodontia = props => {
-  const { classes, items, isLoading, cursos } = props;
-  return (
-    <GridLayout
-      title="Implantodontia"
-      subtitle={`${cursos} cursos`}
-      classes={classes}
-      isLoading={isLoading}
-      showLogo={false}
-    >
-      {items.map(e => (
-        <Grid key={e} item xs={12} md={6} lg={3}>
-          <CarouselCard img={imageFile} />
-        </Grid>
-      ))}
-    </GridLayout>
-  );
-};*/
-
-const Periodontia = props => {
-  const { classes, items, isLoading, cursos } = props;
-  return (
-    <GridLayout
-      title="Periodontia"
-      subtitle={`${cursos} cursos`}
-      classes={classes}
-      isLoading={isLoading}
-      showLogo={false}
-    >
-      {items.map(e => (
-        <Grid key={e} item xs={12} md={6} lg={3}>
-          <CarouselCard img={imageFile} />
-        </Grid>
-      ))}
-    </GridLayout>
-  );
-};
-
-const GridLayout = props => {
-  const { classes, isLoading, title, subtitle, children, showLogo } = props;
-  return (
-    <div className={classes.panel}>
-      <Panel title={title} subtitle={subtitle} showLogo={showLogo}>
-        <Grid container className={classes.root} spacing={16}>
-          {isLoading ? <Loading /> : <>{children}</>}
-        </Grid>
-      </Panel>
-    </div>
-  );
-};
-
 function Home(props) {
   const [lancamentos, setLancamentos] = useState({
     loading: false,
     cursos: 0,
-    items: [1,2,3,4]
+    items: [1,2,3]
   });
   const [periodontia, setPeriodontia] = useState({
     loading: true,
